Move nestListInListItems to blockRendererUtils
